# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Ultima Management API
  version: 0.6.0
  description: |
    Management surface of the Ultima server (a Redis superset clone).
    JWT bearer auth (§9.3) applies to every /api/v1 route except
    /auth/login and /auth/refresh when auth.enabled is set. /metrics is
    guarded by an IP allowlist instead of JWT.
paths:
...
  /api/v1/auth/refresh:
    post:
      operationId: refresh
      summary: Rotate a refresh token (public)
      security: []
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: New token pair (old refresh token is consumed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/auth/password:
    post:
      operationId: changePassword
      summary: Change own password (revokes existing sessions)
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordChangeRequest'
      responses:
        '200':
          description: Password changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/totp/enable:
    post:
      operationId: totpEnable
      summary: Stage a TOTP secret for enrollment
      tags: [auth]
      responses:
        '200':
          description: Enrollment material (confirm with a valid code)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TOTPEnrollment'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /api/v1/auth/totp/confirm:
    post:
      operationId: totpConfirm
      summary: Confirm a staged TOTP secret with a valid code
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TOTPCodeRequest'
      responses:
        '200':
          description: TOTP enabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/auth/totp/disable:
    post:
      operationId: totpDisable
      summary: Disable TOTP (password, plus current code when enabled)
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TOTPDisableRequest'
      responses:
        '200':
          description: TOTP disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/admin/users:
    get:
      operationId: usersList
      summary: List accounts (admin only)
      tags: [admin]
      responses:
        '200':
          description: All accounts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserList'
    post:
      operationId: usersCreate
      summary: Create an account (admin only)
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: Created account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountView'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/admin/users/{name}:
    put:
      operationId: usersUpdate
      summary: Update an account's class/disabled flag (admin only)
      tags: [admin]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: Updated account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountView'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      operationId: usersDelete
      summary: Delete an account (admin only)
      tags: [admin]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/admin/users/{name}/revoke-sessions:
    post:
      operationId: usersRevokeSessions
      summary: Revoke every session of an account (admin only)
      tags: [admin]
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Sessions revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/keys/scan:
    get:
      operationId: scanKeys
      summary: Incremental keyspace scan (SCAN)
      tags: [data]
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
            default: '0'
        - name: match
          in: query
          schema:
            type: string
          description: Glob pattern (MATCH)
        - name: count
          in: query
          schema:
            type: integer
            minimum: 1
          description: Hint for keys per step (COUNT)
        - name: db
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Batch of keys plus the next cursor ("0" = done)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanResult'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/key/{key}:
    get:
      operationId: getKey
      summary: Typed value preview of one key
      tags: [data]
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
        - name: db
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Key type, TTL and a bounded value preview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KeyPreview'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteKey
      summary: Delete one key (DEL)
      tags: [data]
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
        - name: db
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Deletion result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KeyDeleted'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Malformed request body or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing/invalid token, bad credentials, or TOTP required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: State conflict (duplicate account, last admin, job active)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Status:
      type: object
      required: [status]
      properties:
        status:
          type: string
          example: ok
    Error:
      type: object
      required: [status, error]
      properties:
        status:
          type: string
          example: error
        error:
          type: string
...
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
          minLength: 1
          x-oapi-codegen-extra-tags:
            validate: required
    PasswordChangeRequest:
      type: object
      required: [current_password, new_password]
      properties:
        current_password:
          type: string
          minLength: 1
          x-oapi-codegen-extra-tags:
            validate: required
        new_password:
          type: string
          minLength: 1
          x-oapi-codegen-extra-tags:
            validate: required
        totp:
          type: string
    TokenPair:
      type: object
      required: [access_token, refresh_token, expires_in]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
          format: int64
          description: Access-token lifetime in seconds
    TOTPCodeRequest:
      type: object
      required: [totp]
      properties:
        totp:
          type: string
          minLength: 6
          x-oapi-codegen-extra-tags:
            validate: required
    TOTPDisableRequest:
      type: object
      required: [password]
      properties:
        password:
          type: string
          minLength: 1
        totp:
          type: string
    TOTPEnrollment:
      type: object
      required: [secret, provisioning_uri, qr_code_png_base64]
      properties:
        secret:
          type: string
        provisioning_uri:
          type: string
        qr_code_png_base64:
          type: string
    UserCreateRequest:
      type: object
      required: [username, password, class]
      properties:
        username:
          type: string
          minLength: 1
          x-oapi-codegen-extra-tags:
            validate: required
        password:
          type: string
          minLength: 1
          x-oapi-codegen-extra-tags:
            validate: required
        class:
          $ref: '#/components/schemas/AccountClass'
          x-oapi-codegen-extra-tags:
            validate: required,oneof=admin data
    UserUpdateRequest:
      type: object
      properties:
        class:
          $ref: '#/components/schemas/AccountClass'
          x-oapi-codegen-extra-tags:
            validate: "omitempty,oneof=admin data"
        disabled:
          type: boolean
    AccountClass:
      type: string
      enum: [admin, data]
    AccountView:
      type: object
      required: [username, class, totp_enabled, disabled, created_at, updated_at]
      properties:
        username:
          type: string
        class:
          $ref: '#/components/schemas/AccountClass'
        totp_enabled:
          type: boolean
        disabled:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UserList:
      type: object
      required: [users]
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/AccountView'
...
    ScanResult:
      type: object
      required: [cursor, keys]
      properties:
        cursor:
          type: string
        keys:
          type: array
          items:
            type: string
...
    KeyPreview:
      type: object
      required: [key, type, ttl_ms, value]
      properties:
        key:
          type: string
        type:
          type: string
          enum: [string, hash, list, set, zset]
        ttl_ms:
          type: integer
          format: int64
          description: -1 no TTL, -2 no key (never returned; 404 instead)
        value:
          description: >-
            Type-shaped preview, bounded to the first 100 elements for
            collections: string → string, hash → object, list/set → array
            of strings, zset → array of ZSetMember.
    KeyDeleted:
      type: object
      required: [deleted]
      properties:
        deleted:
          type: boolean
security:
  - bearerAuth: []